bug fix: `helpers.readJson` couldn't handle relative paths
used by `DeployConfig.fromFile` which is in turn used by the `new` command
when importing a config. I don't remember why I used `require.resolve`
in that function. It might be a leftover from very early dev when the
deploy config was actually a `.js` module.

Also added a couple of tests.