give the db its own security group
this is necessary because it also needs to have its own removal policy.
if the stack is deleted and the databases are set to RETAIN
cloudformation will fail because it will try to delete the security
group that is still associated with the dbs (which are not getting
deleted). so we give the db it's own security group and set it's removal
policy same as with the parameter groups.